+++ src/bb_strategies.jl
@@ -19,12 +19,12 @@ function branch_mostinfeasible(m, node, disc2var_idx)
 end
 
 """
-init_strong_restart!(node, var_idx, disc_var_idx, l_nd, r_nd, reasonable_disc_vars, infeasible_disc_vars,
+init_strong_restart!(node, opts, var_idx, disc_var_idx, l_nd, r_nd, reasonable_disc_vars, infeasible_disc_vars,
  left_node, right_node, strong_restart, max_gain_var, atol, var_type, disc2var_idx)
 
 Tighten the bounds for the node and check if there are variables that need to be checked for a restart.
 """
-function init_strong_restart!(node, var_idx, disc_var_idx, l_nd, r_nd, 
+function init_strong_restart!(node, opts, var_idx, disc_var_idx, l_nd, r_nd, 
                                 reasonable_disc_vars, infeasible_disc_vars, 
                                 left_node, right_node, strong_restart, max_gain_var,
                                 atol, var_type, disc2var_idx)
@@ -34,7 +34,7 @@ function init_strong_restart!(node, var_idx, disc_var_idx, l_nd, r_nd,
 
     # set the bounds directly for the node
     # also update the best bound and the solution
-    if !state_is_optimal(l_nd.relaxation_state)
+    if !state_is_optimal(l_nd.relaxation_state; allow_almost=opts.allow_almost_solved)
         # the solution shouldn't be updated when the current max_gain_var is then type correct
         if max_gain_var == 0 || !is_type_correct(r_nd.solution[max_gain_var],var_type[max_gain_var],atol)
             # if all variables reasonable_disc_vars are type correct => just continue with the tree
@@ -99,9 +99,9 @@ function branch_strong_on!(m,opts,step_obj,
         end
     end
 
-    function get_current_gains(node, l_nd, r_nd)
-        gain_l = sigma_minus(node, l_nd, node.solution[node.var_idx])
-        gain_r = sigma_plus(node,  r_nd, node.solution[node.var_idx])
+    function get_current_gains(opts, node, l_nd, r_nd)
+        gain_l = sigma_minus(opts, node, l_nd, node.solution[node.var_idx])
+        gain_r = sigma_plus( opts, node,  r_nd, node.solution[node.var_idx])
 
         if isinf(gain_l)
             gain = gain_r
@@ -143,6 +143,7 @@ function branch_strong_on!(m,opts,step_obj,
     max_gain_var = 0
     max_gain_disc_var = 0
 
+    allow_almost = opts.allow_almost_solved
     while restart 
         strong_restarts += 1 # is init with -1
         restart = false
@@ -164,9 +165,11 @@ function branch_strong_on!(m,opts,step_obj,
             opts.debug && (strong_step = new_default_strong_branch_step_obj(var_idx))
                
             # no current restart => we can set max_gain and variable
-            gain_l, gain_r, gain = get_current_gains(node, l_nd, r_nd)
+            gain_l, gain_r, gain = get_current_gains(opts, node, l_nd, r_nd)
 
-            if !state_is_optimal(l_nd.relaxation_state) && !state_is_optimal(r_nd.relaxation_state) && counter == 1
+            
+            if !state_is_optimal(l_nd.relaxation_state; allow_almost=allow_almost) &&
+                !state_is_optimal(r_nd.relaxation_state; allow_almost=allow_almost) && counter == 1
                 # TODO: Might be Error/UserLimit instead of infeasible
                 status = :Infeasible
                 left_node = l_nd
@@ -176,8 +179,9 @@ function branch_strong_on!(m,opts,step_obj,
             end
 
             # check if one part is infeasible => update bounds & restart if strong restart is true
-            if !state_is_optimal(l_nd.relaxation_state) || !state_is_optimal(r_nd.relaxation_state)
-                if !state_is_optimal(l_nd.relaxation_state) && !state_is_optimal(r_nd.relaxation_state)
+            if !state_is_optimal(l_nd.relaxation_state; allow_almost=allow_almost) || 
+                !state_is_optimal(r_nd.relaxation_state; allow_almost=allow_almost)
+                if !state_is_optimal(l_nd.relaxation_state; allow_almost=allow_almost) && !state_is_optimal(r_nd.relaxation_state; allow_almost=allow_almost)
                     # TODO: Might be Error/UserLimit instead of infeasible
                     status = :PartlyInfeasible
                     left_node = l_nd
@@ -185,7 +189,7 @@ function branch_strong_on!(m,opts,step_obj,
                     opts.debug && add_strong_step(strong_step,l_nd,r_nd)
                     break
                 end
-                need_to_resolve, restart, new_infeasible_disc_vars, set_to_last_var = init_strong_restart!(node, var_idx, disc_var_idx, l_nd, r_nd, reasonable_disc_vars, infeasible_disc_vars, left_node, right_node, strong_restart, max_gain_var, opts.atol, m.var_type, disc2var_idx)
+                need_to_resolve, restart, new_infeasible_disc_vars, set_to_last_var = init_strong_restart!(node, opts, var_idx, disc_var_idx, l_nd, r_nd, reasonable_disc_vars, infeasible_disc_vars, left_node, right_node, strong_restart, max_gain_var, opts.atol, m.var_type, disc2var_idx)
                 infeasible_disc_vars = new_infeasible_disc_vars
 
                 # only variables where one branch is infeasible => no restart and break
@@ -196,7 +200,7 @@ function branch_strong_on!(m,opts,step_obj,
                     right_node = r_nd
                     restart = false
                     need_to_resolve = false
-                    gain_l, gain_r, gain = get_current_gains(node, l_nd, r_nd)
+                    gain_l, gain_r, gain = get_current_gains(opts, node, l_nd, r_nd)
                     max_gain = gain
                     set_temp_gains!(gains, gain_l, gain_r, disc_var_idx)
                     opts.debug && add_strong_step(strong_step,l_nd,r_nd)
@@ -394,8 +398,8 @@ function score(q_m, q_p, mu)
     return (1-mu)*minq+mu*maxq
 end
 
-function diff_obj(node, cnode)
-    if state_is_optimal(cnode.relaxation_state)
+function diff_obj(opts, node, cnode)
+    if state_is_optimal(cnode.relaxation_state; allow_almost=opts.allow_almost_solved)
         return abs(node.best_bound - cnode.best_bound)
     else
         return Inf
@@ -404,5 +408,5 @@ end
 
 f_plus(x) = ceil(x)-x
 f_minus(x) = x-floor(x)
-sigma_plus(node,r_nd,x) = diff_obj(node,r_nd)/f_plus(x)
-sigma_minus(node,l_nd,x) = diff_obj(node,l_nd)/f_minus(x)
+sigma_plus(opts, node,r_nd,x) = diff_obj(opts, node,r_nd)/f_plus(x)
+sigma_minus(opts, node,l_nd,x) = diff_obj(opts, node,l_nd)/f_minus(x)