+++ src/types.jl
@@ -38,6 +38,7 @@ mutable struct SolverOptions
     tabu_list_length                    :: Int64
     num_resolve_nlp_feasibility_pump    :: Int64
     mip_solver                          :: Union{Nothing, JuMP.OptimizerFactory}
+    allow_almost_solved                 :: Bool  
     allow_almost_solved_integral        :: Bool  
     registered_functions                :: Union{Nothing, Vector{RegisteredFunction}}